Add VSCode build scripts
After cloning this repo, run `./setup-vscode.sh`.
This will create the VSCode workspace (that you'll open when working on this project), and copy the `.json` files from `vscode-json` to the `.vscode` directory.

Hitting (Command + Shift + P) will invoke the Command Palette -- from which you can invoke a Task.
(for building source and/or programming the device)