Removed the non-ota target. Refactored the ota target to C++.
Formatted everything in my preferred style, and re-wrote in C++.
Created an abstraction for PWM using CRTP so that the vendor-specific code could be replaced.
A similar pattern could be used for I2C, SPI, and UART if other peripherals were to be added to this template project.

The Bluetooth portion is Infineon-specific -- not sure how that can be abstracted.