Update README.md
The original Infineon example has been modified for modern C++.
Also recommending the developer to use VSCode as this repo is set up for that, and also noted that MCUBoot is a submodule in this repo -- and MCUBoot can be built using the scripts provided here.

You need not dig into MCUBoot to set it up (unless of course you're not using a PSoC6 device -- this repo was tested with the CYBLE-416045-EVAL kit) -- the flashmap for PSoC 6 is provided in the root of the repo.